requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
ccxt==4.5.4
pandas==2.2.3
scikit-learn==1.5.2
numpy==1.24.3
requests==2.32.3
tensorflow-cpu==2.13.0
python-dotenv==1.0.1
pymysql==1.1.1